update parsing script
now collects data from the CSV and FASTA and compares for mismatches (previously we mainly used just the FASTA, but the CSV now has more info). 
Fixes a bug in the subsampling (it used to count gaps as sequence!)
Adds state to division label to avoid ambiguous division names.
If division is missing we fall back to state.