update `augur export` syntax
This commit moves to the newer `augur export v1` syntax. This requires augur ~5.4.0 or augur v6. Additionally, the augur-v6-only `export v2` command is added to the snakemake file (and corresponding config file added), but will not be run unless specifically requirested. This will make it easy to switch to this once augur v6 is being used.