Fix and refactor the create_lat_longs script
Since migrating to the NCBI datasets to pull public metadata.tsv, the geolocation fields
have changed. This commit updates the create_lat_longs script to work with the new fields.

From NCBI datasets, the geolocation fields from general to specific are:

* region (e.g. "North America")
* country (e.g. "USA")
* division (e.g. "California")
* location (e.g. "San Francisco")

From the ingest pipeline, derive the 2 letter state abbreviations and populate the 'state' metadata field.
The more specific city or county names are now in the 'location' field.