ingest-to-phylo: Output message for S3 URLs without hash
I noticed that even though files were update in ingest, the cache still
didn't change and phylo did not run.

Traced it back to incorrect S3 URLs leading to the use of
the default hash in the cache. Output message to the GITHUB_STEP_SUMMARY
so that we can immediately see if an S3 file doesn't have a sha256sum.

<https://github.com/nextstrain/WNV/pull/22#issuecomment-2405780818>