Ingest: annotate EF631122 and EF631123 as unknown collection date
EF631122 and EF631123 have collection date annotations of "05-Aug-1931" and "05-Aug-1931" in location "USA: Illinois, Cook County". However, the first initial case of WNV sequenced in USA should be after 1999.

Searching the literature of the submitters:

EF631123 shows up in a phylogenetic tree within the WN02 grouping, and the paper doesn't mention a pre-1999 sample in the USA:

* Amore, G., Bertolotti, L., Hamer, G.L., Kitron, U.D., Walker, E.D., Ruiz, M.O., Brawn, J.D. and Goldberg, T.L., 2010. Multi-year evolutionary dynamics of West Nile virus in suburban Chicago, USA, 2005–2007. _Philosophical Transactions of the Royal Society B: Biological Sciences_, _365_(1548), pp.1871-1878.

I also searched publications:

* Bertolotti, L., Kitron, U.D., Walker, E.D., Ruiz, M.O., Brawn, J.D., Loss, S.R., Hamer, G.L. and Goldberg, T.L., 2008. Fine-scale genetic variation and evolution of West Nile Virus in a transmission “hot spot” in suburban Chicago, USA. _Virology_, _374_(2), pp.381-389.
* Bertolotti, L., Kitron, U. and Goldberg, T.L., 2007. Diversity and evolution of West Nile virus in Illinois and the United States, 2002–2005. _Virology_, _360_(1), pp.143-149.

And concluded to set their collection dates as unknown or "XXXX-XX-XX" rather than potentially skew downstream analysis with a 1931 date.