Use augur subsample
The augur subsample command is built to replace the pattern of multiple
intermediate augur filter calls with simpler Snakemake rules and
configuration.

Note that this is a breaking change and the old configuration will no
longer work. I've updated the top-level config key from "subsampling" to
"subsample" to make the breaking change more obvious, and also to keep
consistent with a pattern of matching top-level key names with rule
names.