Proximity-based sampling
This config introduces the concept of sequential sampling through
proximity-based sampling, which allows a sample to pick sequences not
randomly, but based on genetic distance to sequences in another sample.

It was proposed with a separate schema in the subsampling design doc¹:

    priorities:
      type: proximity
      focus: focal
      num_per_focal: 5

but I think it can be supported by a single new config entry on the same
level as other sampling options (not nested under a new key). Reasons:

- splitting into type and target (focus) would only make sense if there
  are multiple types. In the design doc, there is a "type: titers" but
  that seems to be pathogen specific.
- num_per_focal defines the sample size, which subsample_max_sequences
  and sequences_per_group (N/A without groups) already takes care of.

¹ <https://docs.google.com/document/d/10jrk4OaLhfo6io8WjtZuZFD9BDpr85GtghO1x5GAOM8/edit?tab=t.0#heading=h.liyn1s8gxsux>