Create build-specific config file for all-lineages
The subsampling configuration is specific to the all-lineages build and
not suitable for other builds, so it doesn't make sense to have these in
the default config file. Having them in default config file worked fine
because "subsampling.regions" and "subsampling.force_include" are
overwritten in all other build-specific configs, but this is fragile and
prone to unintentional side effects.

Similar reasoning for all other keys.

This also makes it more obvious in code that all-lineages is the default
build to run when unspecified.