Create build-specific config file for all-lineages
The subsampling configuration is specific to the all-lineages build and
not suitable for other builds, so it doesn't make sense to have these in
the default config file. Having them in default config file worked fine
because "subsample.regions" and "subsample.force_include" are
overwritten in all other build-specific configs, but this is fragile and
prone to unintentional side effects.

I suspect there are other entries in the default config file that should
get a similar treatment, but I'm keeping this PR scoped to just
subsampling-related changes.