Replace "build" wildcard with f-string
Each workflow run only uses a single value for `build`, so it's not
necessary to fill in {build} with Snakemake's wildcard feature. A Python
formatted string replacement is sufficient.

The downside is that this complicates actual usage of wildcards, e.g.

    f"logs/{build}/{{subsample}}/subsampled_strains.txt"

but that is an implementation detail and can be justified by the added
functional improvement that rule names can now be used as targets.