Consolidate all builds into one config / workflow run
The primary motivation is that having 'all-lineages' subsampling
configuration as default and using build-specific config overlays is a
fragile pattern, as it requires the build-specific configs to explicitly
overwrite "subsampling.regions" and "subsampling.force_include".

This means the workflow will now run all builds instead of just
'all-lineages' by default. To run a subset of builds, mark builds to
exclude with '~' (YAML for "null").