Use augur subsample
The augur subsample command is built to replace the pattern of multiple
intermediate augur filter calls with simpler Snakemake rules and
configuration.

Note that this is a breaking change and the old configuration will no
longer work. I've updated the config key from "subsampling" to
"subsample" to make the breaking change more obvious, and also to keep
consistent with a pattern of matching key names with rule names.

The implementation tightly couples the subsample rule with the dump of
config values which was previously only used for debugging. It's not
ideal, but I think it's the best way forward especially when considering
future compatibility with nextstrain run.