Write run time config to a file
This is generally useful for debugging and will be used by augur
subsample.

Note that while this is under the 'rules' folder, it is not a rule.
Instead, it is top-level workflow code run before rules to ensure that
the file is written on every workflow run, including dry runs.