Consolidate all builds into one config / workflow run
The primary motivation is that having 'all-lineages' subsampling
configuration as default and using build-specific config overlays is a
fragile pattern, as it requires the build-specific configs to explicitly
overwrite "subsampling.regions" and "subsampling.force_include".

Even though build names are now provided under 'build_params', listing
them out explicitly via  'builds' is retained to prevent default builds
from leaking into workflow runs using config overlays, and provide an
easy way for CI to run just the all-lineages build.

Basic validation¹ has been added to ensure that 'builds_to_run' is
strictly a subset of available builds. There is more room for validation
such as checking that 'builds_to_run' is defined and as a list, but such
validation is lacking from the rest of the workflow so adding it now
seems out of scope.

¹ error message format taken from augur merge: https://github.com/nextstrain/augur/blob/e8b4d47fd9aae01eb538cc9225992606c7573149/augur/merge.py#L165-L171