Write run config to the same file
The timestamped file did not allow Snakemake to intelligently handle
conditional runs of the subsample rule in the case of config changes in
the subsample block.

In the future, if we really want to store each run's config in its own
timestamped file, that could be written as a second file.