Use shared write_config()
This function is available in shared/vendored/snakemake/config.smk. The
include statement for that file has been moved to before the include
statement for the workflow's config.smk.

The Snakefile would look cleaner if the worfklow's config.smk was moved
further down, however its current placement is required for config
validation.